requirements.txt
keras==2.10.0
matplotlib==3.10.1
numpy==1.25.0
pandas==2.2.3
pillow==10.4.0
scikit-learn==1.6.1
tensorflow==2.10.0
streamlit==1.22.0
protobuf==3.19.6